Renamed Files and started working on error handling
