+++ Vehicle.java
@@ -1,0 +1,6 @@
+// Vehicle class
+
+public abstract class Vehicle {
+    public abstract void displayVariants();
+    public abstract double getPrice(int choice);
+}
